pyproject.toml
[project]
name = "gmail-agent"
version = "0.1.0"
description = "This is a gmail assistant which manages your emails and responds on your behalf "
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.45",
    "langchain-groq>=0.2.5",
    "langgraph>=0.3.11",
    "python-dotenv>=1.0.0",
    "groq>=0.5.0",
    "langchain-ollama>=0.2.3",
    "ipykernel>=6.29.5",
    "google-cloud-bigquery>=3.30.0",
    "sqlalchemy-bigquery>=1.13.0",
    "langchain-openai>=0.3.9",
    "qdrant-client>=1.13.3",
    "sentence-transformers>=4.0.1",
    "bs4>=0.0.2",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.0",
    "dotenv>=0.9.9",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "requests>=2.32.3",
    "google-api-python-client>=2.166.0",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.44.1",
    "google-cloud-firestore>=2.20.2",
]
